refactor: move Banners component to shared components
